ai-engine/requirements.txt
# GitCompass AI Engine - Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# NLP & ML
spacy>=3.7.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0

# Document Processing
pdfminer.six>=20221105
python-docx>=1.1.0

# HTTP Client
aiohttp>=3.9.0
requests>=2.31.0

# Google Gemini AI
google-generativeai>=0.8.0

# Environment
python-dotenv>=1.0.0

# Utilities
pydantic>=2.5.0
